MySQL part is working! Hooray!!!

+++ src/main/kotlin/com/romanhruska/data/mysql/MessagesTable.kt
@@ -1,7 +1,7 @@
 package com.romanhruska.data.mysql
 
 import org.ktorm.schema.Table
-import org.ktorm.schema.int
+import org.ktorm.schema.long
 import org.ktorm.schema.varchar
 
 object MessagesTable: Table<Nothing>("Messages") {
@@ -9,5 +9,5 @@ object MessagesTable: Table<Nothing>("Messages") {
     val id = varchar("id")
     val username = varchar("username")
     val text = varchar("text")
-    val timestamp = int("timestamp")
+    val timestamp = long("timestamp")
 }
